simplify tree strategy, improve and add tests
